refactor string extension
