+++ OLCore/Classes/Extensions/String+Format.swift
@@ -1,0 +1,95 @@
+//
+//  String+Format.swift
+//  OLCore
+//
+//  Created by Sofyan Fradenza Adi on 29/10/19.
+//
+
+import Foundation
+
+extension String {
+    static let numberFormatter = NumberFormatter()
+
+    subscript (i: Int) -> String {
+        if isEmpty || i >= count { return DefaultValue.emptyString }
+        return String(self[index(startIndex, offsetBy: i)])
+    }
+
+    public var doubleValue: Double {
+        String.numberFormatter.decimalSeparator = Separator.decimalEN
+        if let result =  String.numberFormatter.number(from: self) {
+            return result.doubleValue
+        } else {
+            String.numberFormatter.decimalSeparator = Separator.decimalID
+            if let result = String.numberFormatter.number(from: self) {
+                return result.doubleValue
+            }
+        }
+        return 0
+    }
+
+    public var floatValue: Float {
+        String.numberFormatter.decimalSeparator = Separator.decimalEN
+        if let result =  String.numberFormatter.number(from: self) {
+            return result.floatValue
+        } else {
+            String.numberFormatter.decimalSeparator = Separator.decimalID
+            if let result = String.numberFormatter.number(from: self) {
+                return result.floatValue
+            }
+        }
+        return 0
+    }
+
+    public var digits: String {
+        return components(separatedBy: CharacterSet.decimalDigits.inverted)
+            .joined()
+    }
+
+    public func withThousandSeparator() -> String {
+        return Formatter.thousandSeparator.string(for: Int(self.digits)) ?? DefaultValue.emptyString
+    }
+
+    public func formatInIndonesianMobilePhone() -> String {
+        var number = self.digits
+        if self.hasPrefix("+") { number = "+" + number }
+        let prefix = "+" + CountryCode.indonesia
+        if number.hasPrefix(prefix) {
+            number = String(number.dropFirst(prefix.count))
+            number = "0" + number
+        }
+        return number
+    }
+
+    public func toAccessibilityFormat() -> String {
+        return self.replacingOccurrences(
+            of: Separator.whitespace,
+            with: Separator.accessibilityId
+            ).lowercased()
+    }
+
+    public func removeAllWhitespaces() -> String {
+        return String(self.filter{ !" \n\t\r".contains($0) })
+    }
+
+    public func getSuffix(_ maxLength: Int) -> String {
+        var length = maxLength
+        if self.count < maxLength {
+            length = self.count
+        }
+        return String(suffix(length))
+    }
+
+    public func getFirstWord() -> String {
+        return components(separatedBy: DefaultValue.whitespace).first ?? DefaultValue.emptyString
+    }
+
+    public mutating func append(
+        text: String,
+        separator: String = DefaultValue.emptyString
+    ) {
+        if text.isEmpty { return }
+        if !self.isEmpty { self += separator }
+        self += text
+    }
+}
