+++ OLCore/Classes/Extensions/String+Validation.swift
@@ -1,0 +1,26 @@
+//
+//  String+Validation.swift
+//  OLCore
+//
+//  Created by Sofyan Fradenza Adi on 29/10/19.
+//
+
+import Foundation
+
+extension String {
+    public func isValidURL() -> Bool {
+        if let url = NSURL(string: self) {
+            return UIApplication.shared.canOpenURL(url as URL)
+        }
+        return false
+    }
+
+    public func isValid(regexRule: String) -> Bool {
+        return NSPredicate(format: "SELF MATCHES %@", regexRule).evaluate(with: self)
+    }
+
+    public func isBackspace() -> Bool {
+        guard let char = self.cString(using: String.Encoding.utf8) else { return false }
+        return strcmp(char, "\\b") == -92
+    }
+}
